requirements.txt
apache-beam==2.19.0
appdirs==1.4.3
appnope==0.1.0
astroid==2.3.1
attrs==19.3.0
avro-python3==1.9.2.1
backcall==0.1.0
beautifulsoup4==4.9.0
black==19.10b0
cachetools==3.1.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
configparser==5.0.0
crcmod==1.7
cycler==0.10.0
databricks-cli==0.9.1
decorator==4.4.2
dill==0.3.1.1
docopt==0.6.2
fastavro==0.21.24
fasteners==0.15
future==0.18.2
google-api-core==1.16.0
google-apitools==0.5.28
google-auth==1.12.0
google-cloud-bigquery==1.17.1
google-cloud-bigtable==1.0.0
google-cloud-core==1.3.0
google-cloud-datastore==1.7.4
google-cloud-pubsub==1.0.2
google-resumable-media==0.4.1
googleapis-common-protos==1.51.0
grpc-google-iam-v1==0.12.3
grpcio==1.27.2
hdfs==2.5.8
httplib2==0.12.0
idna==2.9
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.16.0
joblib==0.14.1
jupyter-client==6.1.2
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.2
matplotlib==3.2.1
mccabe==0.6.1
mock==2.0.0
monotonic==1.5
nltk==3.4.5
numpy==1.18.2
oauth2client==3.0.0
pandas==1.0.3
parso==0.6.2
pathspec==0.7.0
pbr==5.4.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
prompt-toolkit==3.0.5
protobuf==3.11.3
ptyprocess==0.6.0
pyarrow==0.15.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
Pygments==2.6.1
pylint==2.4.2
pymongo==3.10.1
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
regex==2020.2.20
requests==2.23.0
rope==0.14.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
six==1.14.0
sklearn==0.0
soupsieve==2.0
tabulate==0.8.7
toml==0.10.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
typed-ast==1.4.0
typing==3.7.4.1
typing-extensions==3.7.4.1
urllib3==1.25.8
wcwidth==0.1.9
wordcloud==1.6.0
wrapt==1.11.2
